+++ rxandroidble/src/main/java/com/polidea/rxandroidble/exceptions/BleServiceNotFoundException.java
@@ -16,6 +16,6 @@ public class BleServiceNotFoundException extends BleException {
 
     @Override
     public String toString() {
-        return "BleServiceNotFoundException{" + "serviceUUID=" + serviceUUID + '}';
+        return "BleServiceNotFoundException{serviceUUID=" + serviceUUID + '}';
     }
 }