+++ rxandroidble/src/main/java/com/polidea/rxandroidble/utils/ConnectionSharingAdapter.java
@@ -1,4 +1,4 @@
-package com.polidea.rxandroidble.sample.util;
+package com.polidea.rxandroidble.utils;
 
 import com.polidea.rxandroidble.RxBleConnection;
 
@@ -6,6 +6,30 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import rx.Observable;
 
+/**
+ * Observable transformer that can be used to share connection between many subscribers.
+ * <p>
+ * Example use:
+ * <pre>
+ * UUID characteristicUUID = UUID.fromString("70a5bfcc-a0ec-4091-985e-d5506a31c921");
+ * Observable<RxBleConnection> connectionObservable = bleDevice
+ * .establishConnection(this, false)
+ * .compose(new ConnectionSharingAdapter());
+ *
+ * connectionObservable
+ * .flatMap(rxBleConnection -> rxBleConnection.setupNotification(characteristicUUID))
+ * .flatMap(notificationObservable -> notificationObservable)
+ * .subscribe(bytes -> {
+ * // React on characteristic changes
+ * });
+ *
+ * connectionObservable
+ * .flatMap(rxBleConnection -> rxBleConnection.writeCharacteristic(characteristicUUID, "some text".getBytes()))
+ * .subscribe();
+ * </pre>
+ *
+ * @see com.polidea.rxandroidble.exceptions.BleAlreadyConnectedException
+ */
 public class ConnectionSharingAdapter implements Observable.Transformer<RxBleConnection, RxBleConnection> {
 
     private final AtomicReference<Observable<RxBleConnection>> connectionObservable = new AtomicReference<>();